+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project starting adhering to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) as of version 0.7.0.
 
-## [Unreleased]
+## [1.1.0] - 2019-05-14
+
+- Add a custom Course Feedback tab.  
+- Removed unused views.
 
 ## [1.0.0] - 2019-05-14